Increase precision of player data sending timer

+++ src/game/player/Player.java
@@ -172,7 +172,7 @@ public class Player {
     }
 
 
-    private static float playerDataTimerTicker = 0f;
+    private static double playerDataTimerTicker = 0f;
 
     public static void playersOnTick() {
